+++ .gitattributes
@@ -9,3 +9,5 @@
 *.sh text eol=lf
 *.py text eol=lf
 *.bin binary
+
+*DoAutoCRLF*.approved.* binary diff=text
