+++ GitCommands/Settings/AutoCRLFType.cs
@@ -1,0 +1,12 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace GitCommands.Settings
+{
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    public enum AutoCRLFType
+    {
+        @true,
+        input,
+        @false
+    }
+}
