+++ GitCommands/Settings/SettingsKind.cs
@@ -1,15 +1,5 @@
-using System.Diagnostics.CodeAnalysis;
-
 namespace GitCommands.Settings
 {
-    [SuppressMessage("ReSharper", "InconsistentNaming")]
-    public enum AutoCRLFType
-    {
-        @true,
-        input,
-        @false
-    }
-
     public enum SettingsKind
     {
         /// <summary>
