+++ UnitTests/GitCommands.Tests/Helpers/AutoCrlfExtensionsTests.DoAutoCRLF_should_not_unnecessarily_duplicate_line_ending.ForScenario.MacLines.true.approved.txt
@@ -1,0 +1,4 @@
+﻿Line 1
+Line 2
+
+Line 3
