+++ UnitTests/GitCommands.Tests/Helpers/AutoCrlfExtensionsTests.DoAutoCRLF_should_not_unnecessarily_duplicate_line_ending.ForScenario.UnixLines.input.approved.txt
