+++ UnitTests/GitCommands.Tests/Helpers/MockData/MacLines.bin
