+++ UnitTests/GitCommands.Tests/Helpers/MockData/UnixLines.bin
