+++ UnitTests/GitCommands.Tests/Helpers/MockData/WindowsLines.bin
