--- UnitTests/GitUI.Tests/Editor/FileViewerTextTests.DoAutoCRLF_should_not_unnecessarily_duplicate_line_ending.ForScenario.WindowsLines.input.approved.txt
@@ -1,4 +1,0 @@
-﻿Line 1
-Line 2
-
-Line 3
